Modified code to fix bug with returning wrong base_word
For a word like 'phone,' there is an entry for the -phone combining form, e.g. homophone. The way the code was written it returned the base word -phone rather than phone. Now it returns the first entry, which likely will be the one with the correct form that the user wants.